requirements.txt
numpy==1.20.3
tqdm==4.57.0
matplotlib==3.5.2
Pillow==9.1.1
